Added Study tests, fixing routes and moved functionality - fixed #17 - fixed #18

+++ src/api/services/studies.js
@@ -2,6 +2,7 @@ const ServerError = require('../../lib/error');
 var mongoose = require('mongoose');
 
 var StudiesController = require('../../lib/studiescontroller');
+var UsersController = require('../../lib/userscontroller');
 var GroupsController = require('../../lib/groupscontroller');
 var AllocatorsController = require('../../lib/allocatorscontroller');
 var TestsController = require('../../lib/testscontroller');
@@ -17,7 +18,7 @@ var TestsController = require('../../lib/testscontroller');
 module.exports.getStudies = async (options) => {
   var result = { status: 200, data: {} };
   try{
-    result.data = await StudiesController.getStudies({});
+    result.data = await StudiesController.getStudies({owners: options.user.data.username});
   }catch(e){
     result = { status: 500, data: e };
   }
@@ -62,8 +63,12 @@ module.exports.getStudy = async (options) => {
     if(mongoose.Types.ObjectId.isValid(options.id)){
       var study = await StudiesController.getStudy(options.id);
       if(study !== null){
-        result = { status: 200, data: study };
-      }
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          result = { status: 200, data: study };
+        }else{
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
+        }
+      } 
     }else{
       result = { status: 400, data: {message: 'ObjectId is not valid'} };
     }
@@ -85,7 +90,27 @@ module.exports.updateStudy = async (options) => {
 
   if(mongoose.Types.ObjectId.isValid(options.id)){
     try{
-      await StudiesController.updateStudy(options.id, options.body);
+      var study = await StudiesController.getStudy(options.id);
+      if(study !== null){
+
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          if(options.body.owners.indexOf(options.user.data.username) !== -1){
+
+            let ownersadded = options.body.owners.filter(x => !study.owners.includes(x));
+            var loadedownersadded = await UsersController.getUsers({"username" : {"$in" : ownersadded}});
+
+            if(loadedownersadded.length !== ownersadded.length){
+              result = { status: 404, data: {message: 'An owner added does not exist'} };
+            }else{
+              await StudiesController.updateStudy(options.id, options.body);
+            }
+          }else{
+            result = { status: 400, data: {message: 'Teacher cannot remove itself from the study'} };
+          }
+        }else{
+          result = { status: 401, data: {message: 'User is not authorized to update this study.'} };
+        }
+      }
     }catch(e){
       result = { status: 500, data: e };
     }
@@ -107,7 +132,16 @@ module.exports.deleteStudy = async (options) => {
 
   if(mongoose.Types.ObjectId.isValid(options.id)){
     try{
-      if(!await StudiesController.deleteStudy(options.id)){
+      var study = await StudiesController.getStudy(options.id);
+      if(study !== null){
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          if(!await StudiesController.deleteStudy(options.id)){
+            result = { status: 500, data: { message: 'Error deleting the study' } };
+          }
+        }else{
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
+        }
+      }else{
         result = { status: 404, data: { message: 'Study not found' } };
       }
     }catch(e){
@@ -161,77 +195,21 @@ module.exports.getStudyGroups = async (options) => {
 
   try{
     if(mongoose.Types.ObjectId.isValid(options.id)){
-      console.log(options.id);
       var study = await StudiesController.getStudy(options.id);
       if(study !== null){
-        var groups = await GroupsController.getGroups({"_id" : {"$in" : study.groups}});
-        result = { status: 200, data: groups };
-      }
-    }else{
-      result = { status: 400, data: {message: 'ObjectId is not valid'} };
-    }
-  }catch(e){
-    console.log(e);
-    result =  { status: 500, data: e };
-  }
-
-  return result;
-};
-
-/**
- * @param {Object} options
- * @param {String} options.id The study ID
- * @throws {Error}
- * @return {Promise}
- */
-module.exports.addGroupToStudy = async ( options) => {
-  var result = { status: 404, data: {message: 'Not found'} };
-
-  try{
-    if(mongoose.Types.ObjectId.isValid(options.id)){
-      var study = await StudiesController.getStudy(options.id);
-      if(study !== null){
-        if(study.groups.indexOf(options.body.group) === -1){
-          return await StudiesController.addGroupToStudy(options.id, options.body.group);
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          var groups = await GroupsController.getGroups({"_id" : {"$in" : study.groups}});
+          result = { status: 200, data: groups };
         }else{
-          result = { status: 400, data: {message: 'Group already exists in the study.'}  };
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
         }
+      }else{
+        result = { status: 404, data: { message: 'Study not found' } };
       }
     }else{
       result = { status: 400, data: {message: 'ObjectId is not valid'} };
     }
   }catch(e){
-    console.log(e);
-    result =  { status: 500, data: e };
-  }
-
-  return result;
-};
-
-/**
- * @param {Object} options
- * @param {String} options.id The study ID
- * @throws {Error}
- * @return {Promise}
- */
-module.exports.removeGroupFromStudy = async (options) => {
-  var result = { status: 404, data: {message: 'Not found'} };
-
-  try{
-    if(mongoose.Types.ObjectId.isValid(options.id)){
-      var study = await StudiesController.getStudy(options.id);
-      if(study !== null){
-        if(study.groups.indexOf(options.body.group) !== -1){
-          return await StudiesController.removeGroupFromStudy(options.id, options.body.group);
-        }else{
-          result = { status: 400, data: {message: 'Group does not exists in the study.'}  };
-        }
-      }
-    }else{
-      result = { status: 400, data: {message: 'ObjectId is not valid'} };
-    }
-  }catch(e){
-    console.log(e);
     result =  { status: 500, data: e };
   }
 
@@ -249,17 +227,21 @@ module.exports.getStudyTests = async (options) => {
 
   try{
     if(mongoose.Types.ObjectId.isValid(options.id)){
-      console.log(options.id);
       var study = await StudiesController.getStudy(options.id);
       if(study !== null){
-        var groups = await TestsController.getTests({"_id" : {"$in" : study.tests}});
-        result = { status: 200, data: groups };
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          var groups = await TestsController.getTests({"_id" : {"$in" : study.tests}});
+          result = { status: 200, data: groups };
+        }else{
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
+        }
+      }else{
+        result = { status: 404, data: { message: 'Study not found' } };
       }
     }else{
       result = { status: 400, data: {message: 'ObjectId is not valid'} };
     }
   }catch(e){
-    console.log(e);
     result =  { status: 500, data: e };
   }
 
@@ -279,18 +261,188 @@ module.exports.addTestToStudy = async (options) => {
     if(mongoose.Types.ObjectId.isValid(options.id)){
       var study = await StudiesController.getStudy(options.id);
       if(study !== null){
-        if(study.groups.indexOf(options.body.group) === -1){
+        if(study.owners.indexOf(options.user.data.username) !== -1){
           let test = await StudiesController.addTestToStudy(options.id, options.body);
-          result = { status: 400, data: test };
+          result = { status: 200, data: test };
         }else{
-          result = { status: 400, data: {message: 'Group already exists in the study.'}  };
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
         }
+      }else{
+        result = { status: 404, data: { message: 'Study not found' } };
       }
     }else{
       result = { status: 400, data: {message: 'ObjectId is not valid'} };
     }
   }catch(e){
     console.log(e);
+    result =  { status: 500, data: e };
+  }
+
+  return result;
+};
+
+/**
+ * @param {Object} options
+ * @param {String} options.id The test ID
+ * @throws {Error}
+ * @return {Promise}
+ */
+module.exports.getTest = async (options) => {
+  var result = { status: 404, data: {message: 'Not found'} };
+
+  try{
+    if(mongoose.Types.ObjectId.isValid(options.id) && mongoose.Types.ObjectId.isValid(options.testid)){
+      var study = await StudiesController.getStudy(options.id);
+      if(study !== null){
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          var test = await TestsController.getTest(options.id);
+          if(test !== null){
+            result = { status: 200, data: test };
+          }
+        }else{
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
+        }
+      }else{
+        result = { status: 404, data: { message: 'Study not found' } };
+      }
+    }else{
+      result = { status: 400, data: {message: 'ObjectId is not valid'} };
+    }
+  }catch(e){
+    result =  { status: 500, data: e };
+  }
+
+  return result;
+};
+
+/**
+ * @param {Object} options
+ * @param {String} options.id The study ID
+ * @throws {Error}
+ * @return {Promise}
+ */
+module.exports.updateTest = async (options) => {
+  var result = { status: 200, data: {message: 'Test updated'} };
+
+  if(mongoose.Types.ObjectId.isValid(options.id) && mongoose.Types.ObjectId.isValid(options.testid)){
+    try{
+      await TestsController.updateTest(options.id, options.body);
+    }catch(e){
+      result = { status: 500, data: e };
+    }
+  }else{
+    result = { status: 400, data: { message: 'ObjectId is not valid' } };
+  }
+  
+  return result;
+};
+
+/**
+ * @param {Object} options
+ * @param {String} options.id The study ID
+ * @throws {Error}
+ * @return {Promise}
+ */
+module.exports.deleteTest = async (options) => {
+  var result = { status: 200, data: {message: 'Test deleted'} };
+
+  try{
+    if(mongoose.Types.ObjectId.isValid(options.id) && mongoose.Types.ObjectId.isValid(options.testid)){
+      var study = await StudiesController.getStudy(options.id);
+      if(study !== null){
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          let ntest = study.tests.indexOf(options.testid);
+          if(ntest !== -1){
+            study.tests.splice(ntest,1);
+            if(!await StudiesController.updateStudy(options.id, study)){
+              result = { status: 500, data: 'Error deleting the test' };
+            }
+          }
+        }else{
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
+        }
+      }else{
+        result = { status: 404, data: { message: 'Study not found' } };
+      }
+    }else{
+      result = { status: 400, data: {message: 'ObjectId is not valid'} };
+    }
+  }catch(e){
+    result =  { status: 500, data: e };
+  }
+  
+  return result;
+};
+
+
+/**
+ * @param {Object} options
+ * @param {String} options.id The test ID
+ * @throws {Error}
+ * @return {Promise}
+ */
+module.exports.addActivityToTest = async (options) => {
+  try {
+    options.body.test = options.id;
+
+    let study = await StudiesController.getStudy(options.id);
+    if(!test){
+      return { status: 404, data: { message: 'Study not found' } };
+    }
+
+    let test = await TestsController.getTest(options.testid);
+    if(!test){
+      return { status: 404, data: { message: 'Test not found' } };
+    }
+
+    let activity = ActivitiesController.castToClass(await ActivitiesController.addActivity(options.body));
+
+    test.activities.push(activity.id);
+    await TestsController.updateTest(options.testid, test);
+
+    let participants = await StudiesController.getParticipants(study);
+    await activity.addParticipants(participants);
+
+    return {status: 500, data: activity };
+  }catch(e){
+    console.log(e);
+    return {status: 500, data: e };
+  }
+
+  return { status: 200, data: group };
+};
+
+/**
+ * @param {Object} options
+ * @param {String} options.id The test ID
+ * @throws {Error}
+ * @return {Promise}
+ */
+module.exports.getTestActivities = async (options) => {
+  var result = { status: 404, data: {message: 'Not found'} };
+
+  try{
+    if(mongoose.Types.ObjectId.isValid(options.id) && mongoose.Types.ObjectId.isValid(options.testid)){
+      var study = await StudiesController.getStudy(options.id);
+      if(study !== null){
+        if(study.owners.indexOf(options.user.data.username) !== -1){
+          var test = await TestsController.getTest(options.id);
+          if(test !== null){
+            var activities = await ActivitiesController.getActivities({"_id" : {"$in" : test.activities}});
+            result = { status: 200, data: activities };
+          }else{
+            result = { status: 404, data: { message: 'Test not found' } };
+          }
+        }else{
+          result = { status: 401, data: { message: 'You are not owner of the study' } };
+        }
+      }else{
+        result = { status: 404, data: { message: 'Study not found' } };
+      }
+    }else{
+      result = { status: 400, data: {message: 'ObjectId is not valid'} };
+    }
+  }catch(e){
     result =  { status: 500, data: e };
   }
